Changed hunger from float to long. Now 1 hunger ticks down each millisecond. Working

+++ Consumable.java
@@ -2,5 +2,5 @@
 
 
 interface Consumable{
-    public double consume();    // Returns the amount of nutrition it had in it.
+    public long consume();    // Returns the amount of nutrition it had in it.
 }
